Support optional 'source' in ARP rules
Data in ARP is structured in an associative array:

    attribute name => [ list of value rules ]

We now support storing an optional source for each attribute:

    attribute name => [
        [
            value => '...'
            source => '...'
        ]
    ]

Sources have no effect on the enforcement of the policy. For example,
consider the following hypothetical situation:

  - when an ARP specifies the attribute 'eduPersonOrcid' should be
    included from 'source1'

  - but the aggregator returns no eduPersonOrcid attribute, while the
    IdP does return the attribute

In those cases then the attribute will be included in the SAML response
because the source is not considered when matching attributes.