Allow caching all database queries for the lifetime of a request
This commit introduces a new metadata repository as an alternative to
the regular doctrine metadata repository. It wraps the regular
doctrine metadata repository but caches the result of each distinct
query for the request lifetime. This mechanism is safe to use for
EngineBlock since it doesn't write metadata, and we're not concerned
with metadata updates during an authentication request.

In test cases, using the caching wrapper reduces the number of queries
during a single authentication request from approximately 40 to 15.

In order to use this new repository in EngineBlock, replace the
following line in engineblock.ini:

    metadataRepository.database.type = Doctrine

With:

    metadataRepository.database.type = CachedDoctrine