Delete entities on metadata push
A metadata push by service registry contains metadata of all entities
that engineblock should know about. So all entities in the roles table
not present in the pushed metadata should be deleted on push.

This was already implemented but did not function as intended:

 - the deletion logic did not take into account entities that were
   updated instead of created

 - doctrines 'partial reference' cannot be used to delete entities
   since it does not load the database identifier when loaded by
   entity id

This commit changes the metadata deletion logic by keeping track of
what entity ids should be deleted, and deleting them with one query
when the metadata import has finished.