Support combining the push and pull metadata backends
Using both the push and pull matadata backends did not work as
expected. When fetching ARP or entity manipulations, the first backend
that actually has the entity is used. But when fetching the list of
allowed idps of a service provider, all backends were queried and the
combined results were returned. This did not work for the pull backend
because it throws an error when queried about entities it does not
know about.

This commit fixes the issue by never using all backends, but only the
first backend that actually has the entity.